[anchor-position] Resolve anchor functions in getComputedStyle()
Currently, when getComputedStyle() returns values for `left` etc,
it needs to re-resolve them (after layout) into a pixel value.

This patch provides an NGAnchorEvaluatorImpl for the resolution, so
that anchor functions can also be resolved.

There are also a few cases that are unsolved and left for future.
They are marked with TODOs.

Fixed: 1423493
Change-Id: I319543230008e61a8363fdebb111a9840adc6e5d