[anchor-position] Stop adjusting ComputedStyle for anchor queries
`anchor()` and `anchor-size()` should be invalid on other elements.
This patch changes the ensuring of this to used value time instead
of computed value time, so that they still cascade as normal.

This is ensured by making the property getters protected on `ComputedStyleBase`, which forces all callers (with a few exceptions)
to use the new `ComputedStyle::UsedFoo()` getters.

The exceptions are the new friend classes in `ComputedStyle`:
- `LengthPropertyFunctions`: this class is for animations only,
  and animations should use the original computed value.
- `StyleResolverTest`: for convenience in unit tests

To prevent creating a copy of `Length` object when accessing these
properties, this patch also changes `Length::Auto()` and
Length::None()` to use static instances.

Bug: 1382151
Change-Id: Ie1a8dcfeed92b79161f99fccec8d41b4e4115738