[Critical CH] Test redirects
We need to verify behavior of critical-ch when it comes to redirects.
We need to know that the critical hints are included in the final
page except in the case of a same-origin redirect where the intermediate
and final page both have critical-hints. Any given load only restarts
once per origin so two differing critical hint sets in a single chain
won't both be respected (only the first will be).

Bug: 1428593
Change-Id: Ic0f3469e26828f1196a6f7066acc5eabaea110d2