[CompositeScrollAfterPaint] Consider composited scrolls in CanUpcastWith
In CompositeScrollAfterPaint, the compositing status of a transform
node not only from HasDirectlyCompositedReasons(), but also
composited scroll status determined by PaintArtifactCompositor. This
requires PropertyTreeState::CanUpcastWith also check composited
scroll status when checking if two transform nodes are within the
same compositing boundary.

Also PaintArtifactCompositor needs to handle a case where a scroll
translation is used before the ScrollHitTest chunk. This happens
for negative z-order scrolling contents in a non-stacking-context
scroller. In the case we always composite the scroller to prefer
performance because it's likely there are overlapping interlacing
scrolling and non-scrolling contents. This can also simplify code
because otherwise we would need to find the ScrollHitTest and
compute composited scroll status.

Bug: 1414885
Change-Id: I03411a883dde11e692432da25b87fb49afd7be74