[soft navigations] Ensure detection follows same-document commits
This CL simplifies the soft navigation detection logic and splits the
task that initiates the URL change with the same-document commit that
follows that change. That also enables us to ensure that the soft
navigation detection always followed that same-document commit, which
will simplify our browser-side logic once we start emitting soft
navigation UKM events.

Change-Id: I1ee1abfdb31a69d16f5ff80c4c299cbafae782f7
Bug: 1370937